fix: multi definition issues
When including (directly or indirectly) debug_asserts.hpp in two
different translation units (cpp-files), than each will create
a definition of the function `pasta_assert_impl`, which causes a
multi definition error of that function when linking.
To solve this, we must either mark it with 'static' or 'inline'.
There a different advantages to either.
I decided to go with 'static' which means, that when translating the
function for a translation unit, do not make the symbols available
for any other files.